ci: fix tests
